+++ docs/source/changes.md
@@ -25,6 +25,7 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 - {pull}`536` allows partialed functions to be task functions.
 - {pull}`540` changes the CLI entry-point and allow `pytask.build(tasks=task_func)` as
   the signatures suggested.
+- {pull}`542` refactors the plugin manager.
 
 ## 0.4.4 - 2023-12-04